+++ src/core/Messages.ts
@@ -4,9 +4,17 @@
 import {classToPlain, plainToClass} from "class-transformer";
 import "reflect-metadata";
 
-export class ButtplugMessage {
+export abstract class ButtplugMessage {
 
   constructor(public Id: number) {
+  }
+
+  abstract get SchemaVersion(): number;
+
+  public DowngradeMessage(): ButtplugMessage {
+    return new Error("Message version downgrade required, but not defined for this message type.",
+                     ErrorClass.ERROR_MSG,
+                     this.Id);
   }
 
   /***
@@ -41,14 +49,14 @@ export class ButtplugMessage {
   }
 }
 
-export class ButtplugDeviceMessage extends ButtplugMessage {
+export abstract class ButtplugDeviceMessage extends ButtplugMessage {
   constructor(public DeviceIndex: number,
               public Id: number) {
     super(Id);
   }
 }
 
-export class ButtplugSystemMessage extends ButtplugMessage {
+export abstract class ButtplugSystemMessage extends ButtplugMessage {
   constructor(public Id: number = 0) {
     super(Id);
   }
@@ -58,12 +66,17 @@ export class Ok extends ButtplugSystemMessage {
   constructor(public Id: number) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
+
 }
 
 export class Ping extends ButtplugMessage {
   constructor(public Id: number) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class Test extends ButtplugMessage {
@@ -71,25 +84,34 @@ export class Test extends ButtplugMessage {
               public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export enum ErrorClass {
-    ERROR_UNKNOWN,
-    ERROR_INIT,
-    ERROR_PING,
-    ERROR_MSG,
-    ERROR_DEVICE,
+  ERROR_UNKNOWN,
+  ERROR_INIT,
+  ERROR_PING,
+  ERROR_MSG,
+  ERROR_DEVICE,
 }
 
 export class Error extends ButtplugSystemMessage {
 
-    constructor(public ErrorMessage: string,
-                public ErrorCode: ErrorClass = ErrorClass.ERROR_UNKNOWN,
-                public Id: number = 1) {
+  constructor(public ErrorMessage: string,
+              public ErrorCode: ErrorClass = ErrorClass.ERROR_UNKNOWN,
+              public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
+/***
+ * DeviceInfo Message class from v0 spec
+ *
+ * Uses a string array for messages, instead of a specifications object.
+ */
 export class DeviceInfo {
   constructor(public DeviceIndex: number,
               public DeviceName: string,
@@ -97,18 +119,82 @@ export class DeviceInfo {
   }
 }
 
-export class DeviceList extends ButtplugSystemMessage {
+export class DeviceListVersion0 extends ButtplugSystemMessage {
   constructor(public Devices: DeviceInfo[],
               public Id: number) {
     super();
   }
+
+  public get Type(): string {
+    return "DeviceList";
+  }
+  get SchemaVersion() { return 0; }
 }
 
-export class DeviceAdded extends ButtplugSystemMessage {
+export class DeviceInfoWithSpecifications {
+  constructor(public DeviceIndex: number,
+              public DeviceName: string,
+              public DeviceMessages: object) {
+  }
+}
+
+export class DeviceListVersion1 extends ButtplugSystemMessage {
+  constructor(public Devices: DeviceInfoWithSpecifications[],
+              public Id: number) {
+    super();
+  }
+
+  public get Type(): string {
+    return "DeviceList";
+  }
+
+  public DowngradeMessage(): ButtplugMessage {
+    // This is going to look mostly the same, we just need to reduce our devices
+    // down to use string message lists instead of specification lists.
+    const oldDevices: DeviceInfo[] = [];
+    for (const newDevice of this.Devices) {
+      oldDevices.push(new DeviceInfo(newDevice.DeviceIndex,
+                                     newDevice.DeviceName,
+                                     Object.keys(newDevice.DeviceMessages)));
+    }
+    return new DeviceListVersion0(oldDevices, this.Id);
+  }
+
+  get SchemaVersion() { return 1; }
+}
+
+export class DeviceAddedVersion0 extends ButtplugSystemMessage {
   constructor(public DeviceIndex: number,
               public DeviceName: string,
               public DeviceMessages: string[]) {
     super();
+  }
+
+  public get Type(): string {
+    return "DeviceAdded";
+  }
+  get SchemaVersion() { return 0; }
+}
+
+export class DeviceAddedVersion1 extends ButtplugSystemMessage {
+  constructor(public DeviceIndex: number,
+              public DeviceName: string,
+              public DeviceMessages: object) {
+    super();
+  }
+
+  public get Type(): string {
+    return "DeviceAdded";
+  }
+
+  get SchemaVersion() { return 1; }
+
+  public DowngradeMessage(): ButtplugMessage {
+    // This is going to look mostly the same, we just need to reduce our devices
+    // down to use string message lists instead of specification lists.
+    return new DeviceAddedVersion0(this.DeviceIndex,
+                                   this.DeviceName,
+                                   Object.keys(this.DeviceMessages));
   }
 }
 
@@ -116,30 +202,40 @@ export class DeviceRemoved extends ButtplugSystemMessage {
   constructor(public DeviceIndex: number) {
     super();
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class RequestDeviceList extends ButtplugMessage {
   constructor(public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class StartScanning extends ButtplugMessage {
   constructor(public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class StopScanning extends ButtplugMessage {
-    constructor(public Id: number = 1) {
-        super(Id);
-    }
+  constructor(public Id: number = 1) {
+    super(Id);
+  }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class ScanningFinished extends ButtplugSystemMessage {
-    constructor() {
-        super();
-    }
+  constructor() {
+    super();
+  }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class RequestLog extends ButtplugMessage {
@@ -147,6 +243,8 @@ export class RequestLog extends ButtplugMessage {
               public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class Log extends ButtplugSystemMessage {
@@ -154,12 +252,16 @@ export class Log extends ButtplugSystemMessage {
               public LogMessage: string) {
     super();
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class RequestServerInfo extends ButtplugMessage {
-  constructor(public ClientName: string, public Id: number = 1) {
+  constructor(public ClientName: string, public MessageVersion: number = 0, public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 1; }
 }
 
 export class ServerInfo extends ButtplugSystemMessage {
@@ -172,6 +274,8 @@ export class ServerInfo extends ButtplugSystemMessage {
               public Id: number = 1) {
     super();
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class FleshlightLaunchFW12Cmd extends ButtplugDeviceMessage {
@@ -181,6 +285,8 @@ export class FleshlightLaunchFW12Cmd extends ButtplugDeviceMessage {
               public Id: number = 1) {
     super(DeviceIndex, Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class KiirooCmd extends ButtplugDeviceMessage {
@@ -206,6 +312,8 @@ export class KiirooCmd extends ButtplugDeviceMessage {
       return Math.round(pos);
     }
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class SingleMotorVibrateCmd extends ButtplugDeviceMessage {
@@ -214,6 +322,8 @@ export class SingleMotorVibrateCmd extends ButtplugDeviceMessage {
               public Id: number = 1) {
     super(DeviceIndex, Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class StopDeviceCmd extends ButtplugDeviceMessage {
@@ -221,12 +331,16 @@ export class StopDeviceCmd extends ButtplugDeviceMessage {
               public Id: number = 1) {
     super(DeviceIndex, Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class StopAllDevices extends ButtplugMessage {
   constructor(public Id: number = 1) {
     super(Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class LovenseCmd extends ButtplugDeviceMessage {
@@ -235,6 +349,8 @@ export class LovenseCmd extends ButtplugDeviceMessage {
               public Id: number = 1) {
     super(DeviceIndex, Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
 export class VorzeA10CycloneCmd extends ButtplugDeviceMessage {
@@ -244,29 +360,59 @@ export class VorzeA10CycloneCmd extends ButtplugDeviceMessage {
               public Id: number = 1) {
     super(DeviceIndex, Id);
   }
+
+  get SchemaVersion() { return 0; }
 }
 
-export const Messages = {
-  DeviceAdded,
-  DeviceList,
-  DeviceRemoved,
-  Error,
-  FleshlightLaunchFW12Cmd,
-  KiirooCmd,
-  Log,
-  LovenseCmd,
-  Ok,
-  Ping,
-  RequestDeviceList,
-  RequestLog,
-  RequestServerInfo,
-  ScanningFinished,
-  ServerInfo,
-  SingleMotorVibrateCmd,
-  StartScanning,
-  StopAllDevices,
-  StopDeviceCmd,
-  StopScanning,
-  Test,
-  VorzeA10CycloneCmd,
-};
+export class SpeedSubcommand {
+  constructor(public Index: number,
+              public Speed: number) {
+  }
+}
+
+export class VibrateCmd extends ButtplugDeviceMessage {
+  constructor(public Speeds: SpeedSubcommand[],
+              public DeviceIndex: number = -1,
+              public Id: number = 1) {
+    super(DeviceIndex, Id);
+  }
+
+  get SchemaVersion() { return 1; }
+}
+
+export class RotateSubcommand {
+  constructor(public Index: number,
+              public Speed: number,
+              public Clockwise: boolean) {
+  }
+}
+
+export class RotateCmd extends ButtplugDeviceMessage {
+  constructor(public Rotations: RotateSubcommand[],
+              public DeviceIndex: number = -1,
+              public Id: number = 1) {
+    super(DeviceIndex, Id);
+  }
+
+  get SchemaVersion() { return 1; }
+}
+
+export class VectorSubcommand {
+  constructor(public Index: number,
+              public Position: number,
+              public Duration: number) {
+  }
+}
+
+export class LinearCmd extends ButtplugDeviceMessage {
+  constructor(public Vectors: VectorSubcommand[],
+              public DeviceIndex: number = -1,
+              public Id: number = 1) {
+    super(DeviceIndex, Id);
+  }
+
+  get SchemaVersion() { return 1; }
+}
+
+export { DeviceListVersion1 as DeviceList };
+export { DeviceAddedVersion1 as DeviceAdded };