+++ src/devtools/TestDevice.ts
@@ -1,5 +1,5 @@
-import { ButtplugDevice } from "../server/ButtplugDevice";
-import * as Messages from "../core/Messages";
+import { ButtplugDevice, SingleMotorVibrateCmd, FleshlightLaunchFW12Cmd } from "../index";
+import * as Messages from "../index";
 
 export class TestDevice extends ButtplugDevice {
 
@@ -15,9 +15,11 @@ export class TestDevice extends ButtplugDevice {
     this.MsgFuncs.set(Messages.StopDeviceCmd.name, this.HandleStopDeviceCmd);
     if (shouldVibrate) {
       this.MsgFuncs.set(Messages.SingleMotorVibrateCmd.name, this.HandleSingleMotorVibrateCmd);
+      this.MsgFuncs.set(Messages.VibrateCmd.name, this.HandleVibrateCmd);
     }
     if (shouldLinear) {
       this.MsgFuncs.set(Messages.FleshlightLaunchFW12Cmd.name, this.HandleFleshlightLaunchFW12Cmd);
+      this.MsgFuncs.set(Messages.LinearCmd.name, this.HandleLinearCmd);
     }
   }
 
@@ -29,14 +31,34 @@ export class TestDevice extends ButtplugDevice {
     this._connected = connected;
   }
 
+  public get MessageSpecifications(): object {
+    if (this.MsgFuncs.has(Messages.VibrateCmd.name)) {
+      return {
+        VibrateCmd: { FeatureCount: 1 },
+        SingleMotorVibrateCmd: {},
+        StopDeviceCmd: {},
+      };
+    } else if (this.MsgFuncs.has(Messages.LinearCmd.name)) {
+      return {
+        LinearCmd: { FeatureCount: 1 },
+        FleshlightLaunchFW12Cmd: {},
+        StopDeviceCmd: {},
+      };
+    }
+    return {};
+  }
+
   public Disconnect() {
     this.emit("deviceremoved", this);
   }
 
   private HandleStopDeviceCmd = async (aMsg: Messages.StopDeviceCmd): Promise<Messages.ButtplugMessage> => {
-    this.emit("vibrate", 0);
-    this.emit("linear", { position: this._linearPosition,
-                          speed: this._linearSpeed});
+    if (this.MsgFuncs.has(Messages.VibrateCmd.name)) {
+      this.emit("vibrate", 0);
+    } else if (this.MsgFuncs.has(Messages.LinearCmd.name)) {
+      this.emit("linear", { position: this._linearPosition,
+                            speed: this._linearSpeed});
+    }
     return Promise.resolve(new Messages.Ok(aMsg.Id));
   }
 
@@ -47,6 +69,13 @@ export class TestDevice extends ButtplugDevice {
       return Promise.resolve(new Messages.Ok(aMsg.Id));
     }
 
+  private HandleVibrateCmd =
+    async (aMsg: Messages.VibrateCmd): Promise<Messages.ButtplugMessage> => {
+      return this.HandleSingleMotorVibrateCmd(new SingleMotorVibrateCmd(aMsg.Speeds[0].Speed,
+                                                                        aMsg.DeviceIndex,
+                                                                        aMsg.Id));
+    }
+
   private HandleFleshlightLaunchFW12Cmd =
     async (aMsg: Messages.FleshlightLaunchFW12Cmd): Promise<Messages.ButtplugMessage> => {
       this._linearPosition = aMsg.Position;
@@ -55,4 +84,32 @@ export class TestDevice extends ButtplugDevice {
                             speed: this._linearSpeed });
       return Promise.resolve(new Messages.Ok(aMsg.Id));
     }
+
+  private HandleLinearCmd =
+    async (aMsg: Messages.LinearCmd): Promise<Messages.ButtplugMessage> => {
+      if (aMsg.Vectors.length !== 1) {
+        return new Messages.Error("LinearCmd requires 1 vector for this device.",
+                                  Messages.ErrorClass.ERROR_DEVICE,
+                                  aMsg.Id);
+      }
+      // Move between 5/95, otherwise we'll allow the device to smack into hard
+      // stops because of braindead firmware.
+      const range: number = 90;
+      const vector = aMsg.Vectors[0];
+      const currentPosition = vector.Position * 100;
+      const positionDelta: number = Math.abs(currentPosition - this._linearPosition);
+      let speed: number = Math.floor(25000 * Math.pow(((vector.Duration * 90) / positionDelta), -1.05));
+
+      // Clamp speed on 0 <= x <= 95 so we don't break the launch.
+      speed = Math.min(Math.max(speed, 0), 95);
+
+      const positionGoal = Math.floor(((currentPosition / 99) * range) + ((99 - range) / 2));
+      // We'll set this._lastPosition in FleshlightLaunchFW12Cmd, since
+      // everything kinda funnels to that.
+      return await this.HandleFleshlightLaunchFW12Cmd(new Messages.FleshlightLaunchFW12Cmd(speed,
+                                                                                           positionGoal,
+                                                                                           aMsg.DeviceIndex,
+                                                                                           aMsg.Id));
+    }
+
 }