+++ src/devtools/utils.ts
@@ -1,27 +1,7 @@
-import { ButtplugLogger } from "../core/Logging";
-import { ButtplugClient } from "../client/Client";
-import { ButtplugEmbeddedServerConnector } from "../client/ButtplugEmbeddedServerConnector";
-import { ButtplugServer } from "../server/ButtplugServer";
+import { ButtplugClient, ButtplugEmbeddedServerConnector, ButtplugServer } from "../index";
 import { TestDeviceManager } from "./TestDeviceManager";
-import { TestDeviceManagerPanel } from "./TestDeviceManagerPanel";
-import { LogPanel } from "./LogPanel";
 
-class ButtplugDevToolsLogger extends ButtplugLogger {
-  public static set Logger(logger: ButtplugLogger) {
-    ButtplugLogger.sLogger = logger;
-  }
-}
-
-export function CreateLoggerPanel(logger: ButtplugLogger) {
-  LogPanel.ShowLogPanel(logger);
-}
-
-export function CreateDeviceManagerPanel() {
-  TestDeviceManagerPanel.ShowTestDeviceManagerPanel();
-}
-
-export async function CreateDevToolsClient(logger: ButtplugLogger): Promise<ButtplugClient> {
-  ButtplugDevToolsLogger.Logger = logger;
+export async function CreateDevToolsClient(): Promise<ButtplugClient> {
   const client = new ButtplugClient("Test Client");
   const server = new ButtplugServer("Test Server");
   server.AddDeviceManager(TestDeviceManager.Get());