+++ src/devtools/web/index.web.ts
@@ -1,0 +1,4 @@
+export * from "../index";
+export * from "./LogPanel";
+export * from "./TestDeviceManagerPanel";
+export * from "./utils.web";