+++ src/server/ButtplugDevice.ts
@@ -2,7 +2,7 @@ import * as Messages from "../core/Messages";
 import { EventEmitter } from "events";
 import { IButtplugDevice } from "./IButtplugDevice";
 
-export class ButtplugDevice extends EventEmitter implements IButtplugDevice {
+export abstract class ButtplugDevice extends EventEmitter implements IButtplugDevice {
   protected readonly MsgFuncs: Map<string, (aMsg: Messages.ButtplugMessage) => Promise<Messages.ButtplugMessage>> =
     new Map<string, (aMsg: Messages.ButtplugMessage) => Promise<Messages.ButtplugMessage>>();
 
@@ -10,17 +10,19 @@ export class ButtplugDevice extends EventEmitter implements IButtplugDevice {
     super();
   }
 
+  public abstract get MessageSpecifications(): object;
+
   public get Name() {
     return this._name;
   }
 
-  public GetAllowedMessageTypes = (): string[] => {
-    return Array.from(this.MsgFuncs.keys());
+  public get AllowedMessageTypes(): string[] {
+    return Object.keys(this.MessageSpecifications);
   }
 
   public ParseMessage = async (aMsg: Messages.ButtplugMessage): Promise<Messages.ButtplugMessage> => {
     if (!this.MsgFuncs.has(aMsg.Type)) {
-      return new Messages.Error("${this._name} cannot handle message of type ${aMsg.Type} )",
+      return new Messages.Error(`${this._name} cannot handle message of type ${aMsg.Type}`,
                                 Messages.ErrorClass.ERROR_MSG,
                                 aMsg.Id);
     }