+++ src/server/ButtplugServer.ts
@@ -1,20 +1,17 @@
 import * as Messages from "../core/Messages";
+import { GetSchemaVersion } from "../core/MessageUtils";
 import { DeviceManager } from "./DeviceManager";
 import { EventEmitter } from "events";
 import { ServerMessageHub } from "./ServerMessageHub";
 import { IDeviceSubtypeManager } from "./IDeviceSubtypeManager";
 import { ButtplugLogger, ButtplugLogLevel, LogMessage } from "../core/Logging";
+import { RequestServerInfo } from "../core/Messages";
 
 export class ButtplugServer extends EventEmitter {
 
-  public static CanUseBluetooth = async (): Promise<boolean> => {
-    if (navigator === undefined || !("bluetooth" in navigator)) {
-      return false;
-    }
-    return await navigator.bluetooth.getAvailability();
-  }
-
   // Member: PingTimer?
+  private _clientSchemaVersion: number = -1;
+  private _clientName: string;
   private _deviceManager: DeviceManager;
   private _pingTimedOut: boolean = false;
   private _receivedRequestServerInfo: boolean = false;
@@ -58,6 +55,7 @@ export class ButtplugServer extends EventEmitter {
       } else if (this._outgoingLogLevel === ButtplugLogLevel.Off) {
         this._logger.addListener("log", this.OnLogMessage);
       }
+      this._logger.MaximumEventLogLevel = ButtplugLogLevel[logmsg.LogLevel];
       this._outgoingLogLevel = ButtplugLogLevel[logmsg.LogLevel];
       return new Messages.Ok(logmsg.Id);
     case "Ping":
@@ -65,10 +63,20 @@ export class ButtplugServer extends EventEmitter {
       return new Messages.Ok(aMessage.Id);
     case "RequestServerInfo":
       this._logger.Debug(`Server: RequestServerInfo received.`);
+      const msg = aMessage as RequestServerInfo;
+      if (this._clientSchemaVersion > 1) {
+        // Client automatically disconnects on error message.
+        return new Messages.Error(`Client schema (${this._clientSchemaVersion}) newer than server schema (1). ` +
+                                  "Please upgrade server.",
+                                  Messages.ErrorClass.ERROR_INIT,
+                                  id);
+      }
       this._receivedRequestServerInfo = true;
+      this._clientSchemaVersion = msg.MessageVersion;
+      this._clientName = msg.ClientName;
       // TODO: Figure out how to encode this from the package version?
       // TODO: Figure out how to pull message schema version.
-      return new Messages.ServerInfo(0, 0, 9, 1, this._maxPingTime, this._serverName, id);
+      return new Messages.ServerInfo(0, 0, 0, GetSchemaVersion(), this._maxPingTime, this._serverName, id);
     case "Test":
       this._logger.Debug(`Server: Test received.`);
       const testmsg = aMessage as Messages.Test;
@@ -88,7 +96,20 @@ export class ButtplugServer extends EventEmitter {
     this.OnOutgoingMessage(new Messages.Log(ButtplugLogLevel[aMsg.LogLevel], aMsg.Message));
   }
 
-  private OnOutgoingMessage = (aMsg: Messages.ButtplugMessage) => {
-    this.emit("message", aMsg);
+  private OnOutgoingMessage = (msg: Messages.ButtplugMessage) => {
+    if (msg.constructor.name === "Error") {
+      return msg;
+    }
+    if (this._clientSchemaVersion === -1) {
+      return new Messages.Error("Cannot discern client schema version. Was RequestServerInfo message sent?");
+    }
+    while (msg.SchemaVersion !== this._clientSchemaVersion && msg.SchemaVersion > 0) {
+      msg = msg.DowngradeMessage();
+    }
+    // If there was a conversion problem, log as well as returning an error message.
+    if (msg.constructor.name === "Error") {
+      this._logger.Error((msg as Messages.Error).ErrorMessage);
+    }
+    this.emit("message", msg);
   }
 }