+++ tests/test-messages.ts
@@ -1,5 +1,8 @@
 import * as Messages from "../src/core/Messages";
 import { FromJSON } from "../src/core/MessageUtils";
+import { SetupTestSuite } from "./utils";
+
+SetupTestSuite();
 
 describe("Message", () => {
   it("Converts ok message to json correctly",
@@ -15,14 +18,14 @@ describe("Message", () => {
   it("Converts DeviceList message from json correctly",
      () => {
        // tslint:disable-next-line:max-line-length
-       const jsonStr = '[{"DeviceList":{"Id":2,"Devices": [{"DeviceIndex":0,"DeviceName":"Test","DeviceMessages":["Ok","Ping"]},{"DeviceIndex":1,"DeviceName":"Test1","DeviceMessages":["Ok","Ping"]}]}}]';
+       const jsonStr = '[{"DeviceList":{"Id":2,"Devices": [{"DeviceIndex":0,"DeviceName":"Test","DeviceMessages":{"Ok":{},"Ping":{}}},{"DeviceIndex":1,"DeviceName":"Test1","DeviceMessages":{"Ok":{},"Ping":{}}}]}}]';
        // tslint:disable:max-line-length
-       expect(FromJSON(jsonStr)).toEqual([new Messages.DeviceList([new Messages.DeviceInfo(0, "Test", ["Ok", "Ping"]), new Messages.DeviceInfo(1, "Test1", ["Ok", "Ping"])], 2)]);
+       expect(FromJSON(jsonStr)).toEqual([new Messages.DeviceList([new Messages.DeviceInfoWithSpecifications(0, "Test", {Ok: {}, Ping: {}}), new Messages.DeviceInfoWithSpecifications(1, "Test1", {Ok: {}, Ping: {}})], 2)]);
      });
   it("Converts DeviceAdded message from json correctly",
      () => {
-       const jsonStr = '[{"DeviceAdded":{"Id":0,"DeviceIndex":0,"DeviceName":"Test","DeviceMessages":["Ok","Ping"]}}]';
-       expect(FromJSON(jsonStr)).toEqual([new Messages.DeviceAdded(0, "Test", ["Ok", "Ping"])]);
+       const jsonStr = '[{"DeviceAdded":{"Id":0,"DeviceIndex":0,"DeviceName":"Test","DeviceMessages":{"Ok":{},"Ping":{}}}}]';
+       expect(FromJSON(jsonStr)).toEqual([new Messages.DeviceAdded(0, "Test", {Ok: {}, Ping: {}})]);
      });
   it("Converts Error message from json correctly",
      () => {
@@ -58,4 +61,18 @@ describe("Message", () => {
 
        expect(msg2.toJSON()).toEqual('{"KiirooCmd":{"Id":2,"DeviceIndex":3,"Command":"4"}}');
      });
+  it("Handles Device Commands with Subcommand arrays correctly",
+     () => {
+       const jsonStr = '[{"VibrateCmd":{"Id":2, "DeviceIndex": 3, "Speeds": [{ "Index": 0, "Speed": 100}, {"Index": 1, "Speed": 50}]}}]';
+       expect(FromJSON(jsonStr)).toEqual([new Messages.VibrateCmd([{Index: 0, Speed: 100}, {Index: 1, Speed: 50}], 3, 2)]);
+     });
+
+  it("Handles RequestServerInfo correctly across multiple schema versions",
+     () => {
+       const jsonV0Str = '[{"RequestServerInfo":{"Id":2,"ClientName":"TestClient"}}]';
+       expect(FromJSON(jsonV0Str)).toEqual([new Messages.RequestServerInfo("TestClient", 0, 2)]);
+       const jsonV1Str = '[{"RequestServerInfo":{"Id":2,"MessageVersion":1,"ClientName":"TestClient"}}]';
+       expect(FromJSON(jsonV1Str)).toEqual([new Messages.RequestServerInfo("TestClient", 1, 2)]);
+     });
+
 });